完成监测，计费统计，报表功能

+++ src/main/java/cn/edu/bupt/cac/entity/ReportItem.java
@@ -4,6 +4,8 @@ import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 // 请求类
 public class ReportItem {
@@ -15,8 +17,8 @@ public class ReportItem {
     private String duration; // 请求时长，等于结束时间减去开始时间
     private String startTime; // 开始时间
     private String endTime; // 结束时间
-    private double startTemp; // 开始温度
-    private double endTemp; // 结束温度
-    private double energy; // 消耗能量，与请求时长和风速有关
-    private double cost; // 支付费用，与消耗能量有关
+    private BigDecimal startTemp;
+    private BigDecimal endTemp;
+    private BigDecimal energy; // 消耗能量，与请求时长和风速有关
+    private BigDecimal cost; // 支付费用，与消耗能量有关
 }
